Project Overview :
In the Tic-Tac-Toe game, you will see the game's approach is implemented. In this game, two players can play, and a print board having numbers from 1 to 9 will be displayed on the screen, or you can say it box number. You have to choose X or O for the specific box number. For example, if you have to select any number, then X or O will be shown on the print board, and turn for next will be there. The task is to create a Java program to implement a 3×3 Tic-Tac-Toe game for two players.